requirements.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
numpy
